requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.2.5
attrs==19.2.0
audioread==2.1.8
awscli==1.16.306
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
botocore==1.13.42
browserhistory==0.1.2
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
comtypes==1.1.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
djangorestframework==3.10.3
docutils==0.15.2
entrypoints==0.3
firebase-admin==3.0.0
Flask==1.1.1
future==0.17.1
gast==0.2.2
gcloud==0.17.0
get==2019.4.13
gevent==1.4.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.13.1
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
librosa==0.7.0
llvmlite==0.29.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
MouseInfo==0.0.4
msgpack==0.6.2
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numba==0.45.1
numpy==1.17.2
oauth2client==3.0.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0rc1
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyAudio==0.2.11
PyAutoGUI==0.9.47
pycparser==2.19
pycryptodome==3.4.3
PyGetWindow==0.0.7
Pygments==2.4.2
pylint==2.3.1
PyMsgBox==1.0.7
pynmea==0.6.0
pyparsing==2.4.2
pyperclip==1.7.0
Pyrebase==3.0.27
PyRect==0.1.4
pyrsistent==0.15.4
PyScreeze==0.1.22
python-dateutil==2.8.0
python-jwt==2.0.1
python-speech-features==0.6
pyttsx==1.1
pyttsx3==2.81
PyTweening==1.0.3
pytz==2019.2
pywin32==225
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.7.0
resampy==0.2.2
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
serial==0.0.97
six==1.12.0
SoundFile==0.10.2
soupsieve==1.9.4
sqlparse==0.3.0
tb-nightly==1.15.0a20190806
tensorboard==2.0.0
tensorflow==2.0.0rc1
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019080601
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
win32core==221.36
win32gui==221.6
wrapt==1.11.2
zipp==0.6.0
